AR-11719
(cherry picked from commit 27beec2f1523444ce7357abe19a47b3da5dc9bc1)

+++ i21Database/dbo/Stored Procedures/uspCTGetContractPrice.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROCEDURE [dbo].[uspCTGetContractPrice]
 	@intContractHeaderId int
 	,@intContractDetailId int
-	,@dblQuantityToPrice int
+	,@dblQuantityToPrice numeric(18,6)
 	,@strScreen nvarchar(100) /*screen that calls the procedure like "Ticket", "Inventory", etc..*/
 	
 AS
